With new doc version

+++ src/get_vaccine_action.jl
@@ -10,13 +10,8 @@ reach after a horizont time t_horizon a coverage X_C.
 - `t::Float`: time
 - `parameters::DataFrame`: current parameters.
 ...
-    
 """
-
-function get_vaccine_action!(X_C::Float64, t::Float64, parameters)
-    # to be implemented
-    # function get_vaccine_action!(X_C::Float64, t::Float64, parameters, des)
-    #
+function get_vaccine_action!(X_C, t, parameters)
     id = get_stencil_projection(t, parameters)
     t_initial_interval = parameters.t_delivery[id-1]
     t_horizon = t - t_initial_interval
